uadk: fix cleancode issues
1. Use 'intptr_t' or 'uintptr_t' to convert pointer into integer.
2. Forbidden exit function 'pthread_exit' is called.
3. Use 1 blank space(' ') instead of TAB('\t') between the right
   comment and the previous code.
4. Do not add blank lines on the start of a code block defined by braces.

+++ include/wd.h
@@ -201,7 +201,7 @@ static inline void *WD_ERR_PTR(uintptr_t error)
 
 static inline long WD_PTR_ERR(const void *ptr)
 {
-	return (long)ptr;
+	return (intptr_t)ptr;
 }
 
 /**
